[#424] Make drivers tree selector as a reusable component
